updated macos binaries, fixed "save dialog" bug on newer macos versions

+++ renderer/glviewwidget.cpp
@@ -1146,9 +1146,9 @@ void glViewWidget::initTextures()
 		{
 			//float col = std::min(fabs(i-RASTER_SIZE/2.f), fabs(j-RASTER_SIZE/2.f))/(RASTER_SIZE/2.f);
 			double major = std::min(fabs(i-RASTER_SIZE/2.f), fabs(j-RASTER_SIZE/2.f))/(RASTER_SIZE/2.f);
-			double minor = std::min(std::min(fabs(((i+(RASTER_SIZE/20))%(RASTER_SIZE/10)-(RASTER_SIZE/20.f))), fabs(((j+(RASTER_SIZE/20))%(RASTER_SIZE/10)-(RASTER_SIZE/20.f)))), 1.);
+            double minor = std::min((double)std::min(fabs(((i+(RASTER_SIZE/20))%(RASTER_SIZE/10)-(RASTER_SIZE/20.f))), fabs(((j+(RASTER_SIZE/20))%(RASTER_SIZE/10)-(RASTER_SIZE/20.f)))), 1.);
 			major = 1-major;
-			minor = 1-minor;
+            minor = 1-minor;
 			for(int k = 0; k < 9; ++k)
 			{
 				major *= major;
